refactor: 인터페이스 및 레포지토리 파일명 변경
