+++ src/v1/apis/auth/auth.service.ts
@@ -9,7 +9,7 @@ import {
   signupResponseSchema,
 } from './auth.schema.js';
 import { STATUS } from '../../common/constants/status.js';
-import { UserRepositoryInterface } from '../../repositories/persistent/interfaces/user.repository.interface.js';
+import { UserRepositoryInterface } from '../../repositories/persistent/interfaces/user.interface.js';
 import { NotFoundException } from '../../common/exceptions/core.error.js';
 import { FastifyBaseLogger } from 'fastify';
