+++ src/v1/repositories/persistent/interfaces/base.interface.ts
