+++ src/v1/repositories/persistent/interfaces/user.interface.ts
@@ -1,5 +1,5 @@
 import { Prisma, User } from '@prisma/client';
-import { BaseRepository } from './base.repository.interface.js';
+import { BaseRepository } from './base.interface.js';
 
 export interface UserRepository
   extends BaseRepository<User, Prisma.UserCreateInput, Prisma.UserUpdateInput> {
